supervisor/requirements.txt
Flask==2.2.5
docker==6.1.3
pyyaml==6.0
gunicorn==21.2.0
